# inventory-management-system-openapi3_1.yaml
openapi: 3.1.0
info:
  title: Inventory Management System - OpenAPI 3.1
  description: |-
    This is a inventory management system server based on the OpenAPI 3.1 specification. here i am starting with the design first approach in mind! with i can improve the API in general and expose some of the new features in OAS3.
  contact:
    email: 1221344@student.birzeit.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://localhost:9090/ims-api/v1
...
paths:
  /stocks:
    get:
      tags:
        - Stock
      summary: Gets all stocks in the system
      description: Gets all stocks in the system
      operationId: getAllStocks
      responses:
        '200':
          description: The stocks was successfully acquired.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '403':
          description: The client doesn't have access to the stocks
        '404':
          description: Stocks not found.
        '410':
          description: The stocks are no longer available
        '500':
          description: The server encountered an unexpected error while processing the stocks.
        '504':
          description: Unable to process the stocks in time
      security:
        - api_key: []
    post:
      tags: 
        - Stock
      summary: Adds a new stock to the system
      description: Adds a new stock to the system
      operationId: createStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStock'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '400':
          description: Invaild stock data provided
        '401':
          description: The client is unauthorized. the stock was not created
        '409':
          description: The stock provided will conflict the resources on the server
        '415':
          description: Content-type of the request is not supported by this server
        '500':
          description: The server encountered an unexpected error while creating the stock
  /stocks/{id}:
    get:
      tags:
        - Stock
      summary: Gets a stock by id
      description: stock by id
      operationId: getStockById
      parameters:
        - name: id
          in: path
          description: ID of stock to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The stock was successfully acquired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '403':
          description: The client doesn't have access to this stock
        '404':
          description:  Stock with spicified id not found
        '410':
          description: The stock with spicified id are no longer available
        '500':
          description: The server encountered while processing the stock
        '504':
          description: Unable to process the spicified stock in time
    put:
      tags:
        - Stock
      summary: Updates a stock by id
      description: update stock by id
      operationId: updateStockById
      parameters:
        - name: id
          in: path
          description: ID of stock to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStock'
      responses:
        '200':
          description: The stock was successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '400':
          description: Invalid stock id
        '401':
          description:  The client is unauthorized. the stock was not updated
        '404':
          description: Stock with given id is not found
        '409':
          description: The new stock data provided will conflict the resources on the server
        '500':
          description: The server encountered an unexpected error while updating the stock.

    delete:
      tags:
        - Stock
      summary: Deletes a stock by id
      description: Deletes stock by id
      operationId: deleteStockById
      parameters:
        - name: id
          in: path
          description: ID of stock to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The stock was successfully deleted
          content:
            application/json:
              schema:
                type: string
                examples: [The stock was successfully deleted]
        '400':
          description: Invalid stock id
        '401':
          description:  The client is unauthorized. the stock was not deleted
        '403':
          description: The client doesn't have access to spicified stock. Was unable to deleted
        '404':
          description: Stock Not Found
        '500':
          description: The server encountered an unexpected error while deleting the stock
  /stocksByTag/{tag}:
    get:
      tags:
        - Stock
      summary: Gets a stock by Tag
      description: stock by tag
      operationId: getStockByTag
      parameters:
        - name: tag
          in: path
          description: tag of stocks to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [reuseable]
      responses:
        '200':
          description: The stocks with spicified tag were successfully acquired
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '403':
          description: The client doesn't have access to spicified stocks
        '404':
          description: there are no stock with given tag
        '410':
          description: The stocks with spicified tag are no longer available
        '500':
          description: The server encountered while processing the stocks
        '504':
          description: Unable to process the spicified stocks in time
  /stocksByStatus/{status}:
    get:
      tags:
        - Stock
      summary: Gets a stock by status
      description: stock by status
      operationId: getStockByStatus
      parameters:
        - name: status
          in: path
          description: status of stocks to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [available]
      responses:
        '200':
          description:  The stocks with spicified status were successfully acquired
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '403':
          description: The client doesn't have access to spicified stocks
        '404':
          description: there are no stock with given status
        '410':
          description: The stocks with spicified status are no longer available
        '500':
          description: The server encountered an unexpected error while processing the stocks
        '504':
          description: Unable to process the spicified stocks in time
...
  /warehouses/{id}/linkStock/{stock_id}:
    put:
      tags:
        - Warehouse
      summary: Link a stock to spicific warehouse
      description: Link a stock to spicific warehouse
      operationId: linkStock
      parameters:
        - name: id
          in: path
          description: ID of warehouse to link
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
        - name: stock_id
          in: path
          description: ID of stock to link
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The stock was successfully linked with the warehouse
          content:
            application/json:
              schema:
                type: string
                examples: [The stock was successfully linked with the warehouse]
        '400':
          description: Invalid warehouse or stock id
        '401':
          description: The client is unauthorized. the warehouse was not linked with the stock
        '404':
          description: Warehouse or Stock with spicified id not found
        '409':
          description: The link will conflict the resources on the server
        '500':
          description: The server encountered an unexpected error while linking the warehouse
  /providers:
    get:
      tags:
        - Provider
      summary: Gets all providers in the system
      description: Gets all providers in the system
      operationId: getAllProviders
      responses:
        '200':
          description: The providers were successfully acquired
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'
        '403':
          description: The client doesn't have access to the providers
        '404':
          description: there are no providers
        '410':
          description: The providers are no longer available
        '500':
          description: The server encountered an unexpected error while  processing the providers
        '504':
          description:  Unable to process the providers in time
    post:
      tags:
        - Provider
      summary: Adds new provider to the system
      description: Adds new provider to the system
      operationId: createProvider
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProvider'
      responses:
        '201':
          description: The provider was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          description: Invaild data for the provider
        '401':
          description: The client is unauthorized. the provider was not created
        '409':
          description: The given provider will conflict the resources on the server
        '415':
          description: Content-type of the request is not supported by this server
        '500':
          description: The server encountered an unexpected error while creating the provider
  /providers/{id}:
    get:
      tags:
        - Provider
      summary: Gets a provider by id
      description: provider by id
      operationId: getProviderById
      parameters:
        - name: id
          in: path
          description: ID of provider to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The provider was successfully acquired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '403':
          description: The client doesn't have access to this provider
        '404':
          description: there are no provider with given id
        '410':
          description: The provider with spicified id are no longer available
        '500':
          description: The server encountered an unexpected error while processing the provider
        '504':
          description: Unable to process the spicified provider in time
    put:
      tags:
        - Provider
      summary: Updates a provider by id
      description: update provider by id
      operationId: updateProviderById
      parameters:
        - name: id
          in: path
          description: ID of provider to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProvider'
      responses:
        '200':
          description: The provider was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewProvider'
        '400':
          description: Invalid provider id
        '401':
          description: The client is unauthorized. the provider was not updated
        '404':
          description: Provider Not Found
        '409':
          description: The new provider data will conflict the resources on the server
        '500':
          description: The server encountered an unexpected error while updating the provider
    delete:
      tags:
        - Provider
      summary: Deletes a provider by id
      description: Deletes provider by id
      operationId: deleteProviderById
      parameters:
        - name: id
          in: path
          description: ID of provider to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The provider was successfully deleted
          content:
            application/json:
              schema:
                type: string
                examples: ["The provider was successfully deleted"]
        '400':
          description: Invalid provider id
        '401':
          description: The client is unauthorized. the provider was not deleted
        '403':
          description: The client doesn't have access to spicified provider. Was unable to deleted
        '404':
          description: Provider Not Found
        '500':
          description: The server encountered an unexpected error while deleting the provider
  /providers/{id}/linkStock/{stock_id}:
    put:
      tags:
        - Provider
      summary: Link a stock to spicific warehouse
      description: Link a stock to spicific warehouse
      operationId: linkStockProvider
      parameters:
        - name: id
          in: path
          description: ID of provider to link
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
        - name: stock_id
          in: path
          description: ID of stock to link
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The stock was successfully linked with the warehouse
          content:
            application/json:
              schema:
                type: string
                examples: [The stock was successfully linked with the provider]
        '400':
          description: Invalid warehouse or stock id
        '401':
          description: The client is unauthorized. the warehouse was not linked with the stock
        '404':
          description: Warehouse or Stock with spicified id not found
        '409':
          description: The link will conflict the resources on the server
        '500':
          description: The server encountered an unexpected error while linking the warehouse
...
components:
  schemas:
    Stock:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          examples: [1]
        name:
          type: string
          examples: ["bag"]
        price:
          type: number
          examples: [12.2, 303]
        status:
          type: string
          examples: [available]
          enum:
            - available
            - unavailable
            - pending
        tags:
          type: array
          items:
            type: string
          minItems: 1
          examples: [["reuseable"]]
        availableQuantity:
          type: integer
          format: int64
          examples: [12]
    NewStock:
      type: object
      properties:
        name:
          type: string
          examples: ["bag"]
        price:
          type: number
          examples: [12.2, 303]
        status:
          type: string
          examples: [AVAILABLE]
          enum:
            - AVAILABLE
            - RESERVED
            - OUT_OF_STOCK
            - DAMAGED
            - DISCONTINUED
        tags:
          type: array
          items:
            type: string
          minItems: 1
          examples: [["reuseable"]]
        availableQuantity:
          type: integer
          format: int64
          examples: [12]
...
    Provider:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          examples: [1]
        name: 
          type: string
          examples: [alnoor]
        type:
          type: string
          examples: [Person, Company]
    NewProvider:
      type: object
      properties:
        name: 
          type: string
          examples: [alnoor]
        type:
          type: string
          examples: [Person, Company]
...